requirements.txt
cffi==1.16.0
cryptography==41.0.7
dictdiffer==0.9.0
grpcio==1.60.0
protobuf==4.25.1
psycopg2-binary==2.9.9
pycparser==2.21
pygnmi==0.8.13
robotframework==6.1.1
